pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-reader"
version = "0.1.0"
description = "CLI tool for extracting stable txt from academic PDFs for AI reading workflows."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "PyMuPDF>=1.24.0",
  "pdfplumber>=0.11.0",
  "pypdf>=4.2.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]
ocr = [
  "pytesseract>=0.3.10",
  "Pillow>=10.0.0",
]

[project.scripts]
paper-reader = "paper_reader.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
